requirements/base.txt
-i https://pypi.org/simple
aiofiles==23.2.1; python_version >= '3.7'
aiogram==3.3.0; python_version >= '3.8'
aiohttp==3.9.3; python_version >= '3.8'
aiohttp-socks==0.8.4
aiosignal==1.3.1; python_version >= '3.7'
amqp==5.2.0; python_version >= '3.6'
aniso8601==9.0.1
annotated-types==0.6.0; python_version >= '3.8'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.2.0; python_version >= '3.7'
billiard==4.2.0; python_version >= '3.7'
celery==5.3.6; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
channels==4.0.0; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cron-descriptor==1.4.3
django==4.2.9; python_version >= '3.8'
django-axes==6.3.0; python_version >= '3.7'
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cleanup==8.0.0
django-cors-headers==4.3.1; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-filter==23.5; python_version >= '3.7'
django-graphql-jwt==0.4.0; python_version >= '3.6'
django-redis==5.4.0; python_version >= '3.6'
django-timezone-field==6.1.0; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-spectacular==0.27.1; python_version >= '3.6'
flower==2.0.1; python_version >= '3.7'
frozenlist==1.4.1; python_version >= '3.8'
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
humanize==4.9.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kombu==5.3.5; python_version >= '3.8'
magic-filter==1.0.12; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
pillow==10.1.0; python_version >= '3.8'
prometheus-client==0.19.0; python_version >= '3.8'
promise==2.3
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
pycparser==2.21
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-crontab==3.0.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-socks[asyncio]==2.4.4
pytz==2023.4
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.1; python_version >= '3.7'
referencing==0.33.0; python_version >= '3.8'
rpds-py==0.17.1; python_version >= '3.8'
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlparse==0.4.4; python_version >= '3.5'
text-unidecode==1.3
tornado==6.4; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2023.4; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
whitenoise==6.6.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
